+++ exporter/opentelemetry-exporter-otlp/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+- Add meter reference to observers
+  ([#1425](https://github.com/open-telemetry/opentelemetry-python/pull/1425))
+
+## Version 0.16b0
+
+Released 2020-11-25
+
+- Add instrumentation library name and version to OTLP exported metrics
+  ([#1418](https://github.com/open-telemetry/opentelemetry-python/pull/1418))
 - Change temporality for Counter and UpDownCounter
   ([#1384](https://github.com/open-telemetry/opentelemetry-python/pull/1384))
 - Add Gzip compression for exporter